support for adapters
