+++ examples/nextjs/.env
@@ -1,2 +1,2 @@
-SECRET=sause
+SECRET=33
 NEXT_PUBLIC_GREETING=hello
