+++ examples/nextjs/package.json
@@ -14,6 +14,7 @@
     "@types/react": "18.0.38",
     "@types/react-dom": "18.0.11",
     "eslint": "^8.39.0",
+    "joi": "^17.9.2",
     "next": "^13.4.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
