+++ examples/nuxt/.env
@@ -1,2 +1,2 @@
-SECRET=sause
+SECRET=5
 NUXT_PUBLIC_GREETING=hello
