+++ examples/nuxt/env.ts
@@ -1,11 +1,11 @@
-import { createEnv } from "@t3-oss/env-nuxt";
-import { z } from "zod";
+import { joiAdapter, zodAdapter } from "@t3-oss/env-nuxt";
+import Joi from "joi";
 
-export const env = createEnv({
-  server: {
-    SECRET: z.string(),
-  },
+export const env = joiAdapter({
   client: {
-    NUXT_PUBLIC_GREETING: z.string(),
+    NUXT_PUBLIC_GREETING: Joi.string(),
+  },
+  server: {
+    SECRET: Joi.number(),
   },
 });
