+++ examples/nuxt/package.json
@@ -5,8 +5,7 @@
     "build": "nuxt build",
     "dev": "nuxt dev",
     "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "postinstall": "turbo build --filter env-nuxt && nuxt prepare"
+    "preview": "nuxt preview"
   },
   "devDependencies": {
     "@types/node": "^18.15.13",
@@ -14,6 +13,7 @@
   },
   "dependencies": {
     "@t3-oss/env-nuxt": "0.6.0",
+    "joi": "^17.9.2",
     "zod": "^3.20.2"
   }
 }
